+++ src/MissingPersonsRegister.py
@@ -66,10 +66,11 @@ class MissingPersonsRegister(Dataset):
 
     @Dataset.measure_execution_time
     def __clear_collection(self):
-        missing_persons_col = self.db['MissingPersons']
-        count_deleted_documents = missing_persons_col.delete_many({})
-        logging.warning('%s documents deleted. The missing persons collection is empty.', str(
-            count_deleted_documents.deleted_count))
+        if Dataset.is_collection_exists('MissingPersons'):
+            missing_persons_col = self.db['MissingPersons']
+            count_deleted_documents = missing_persons_col.delete_many({})
+            logging.warning('%s documents deleted. The missing persons collection is empty.', str(
+                count_deleted_documents.deleted_count))
 
     @Dataset.measure_execution_time
     def __create_service_json(self):
@@ -99,9 +100,8 @@ class MissingPersonsRegister(Dataset):
 
     @Dataset.measure_execution_time
     def __update_metadata(self):
-        collections_list = self.db.list_collection_names()
         # update or create MissingPersonsRegisterServiceJson
-        if ('ServiceCollection' in collections_list) and (self.serviceCol.count_documents({'_id': 1}, limit=1) != 0):
+        if (Dataset.is_collection_exists('ServiceCollection')) and (self.serviceCol.count_documents({'_id': 1}, limit=1) != 0):
             self.__update_service_json()
             logging.info('MissingPersonsRegisterServiceJson updated')
         else:
@@ -110,10 +110,11 @@ class MissingPersonsRegister(Dataset):
 
     @Dataset.measure_execution_time
     def __delete_collection_index(self):
-        missing_persons_col = self.db['MissingPersons']
-        if 'full_text' in missing_persons_col.index_information():
-            missing_persons_col.drop_index('full_text')
-            logging.warning('Missing persons Text index deleted')
+        if Dataset.is_collection_exists('MissingPersons'):
+            missing_persons_col = self.db['MissingPersons']
+            if 'full_text' in missing_persons_col.index_information():
+                missing_persons_col.drop_index('full_text')
+                logging.warning('Missing persons Text index deleted')
 
     @Dataset.measure_execution_time
     def __create_collection_index(self):
@@ -148,7 +149,7 @@ class MissingPersonsRegister(Dataset):
                                           '{:.10}'.format(result['BIRTH_DATE']), result['LOST_PLACE'],
                                           '{:.10}'.format(result['LOST_DATE'])])
                 print(result_table.get_string(
-                    title='The missing persons register: ' + str(result_count) + ' records found'))
+                    title=f'The missing persons register: {result_count} records found'))
                 logging.warning('The missing persons register: %s records found', str(result_count))
                 print('Only 10 first search results showed')
                 # save all search results into HTML
